web-bundle/subresource-loading/csp-blocked.https.tentative.html

<!DOCTYPE html>
<title>CSP for subresource WebBundle (blocked cases)</title>
<link
  rel="help"
  href="https://github.com/WICG/webpackage/blob/master/explainers/subresource-loading.md"
/>
<meta
  http-equiv="Content-Security-Policy"
  content="
    script-src
      urn:
      https://web-platform.test:8444/resources/testharness.js
      https://web-platform.test:8444/resources/testharnessreport.js
      'unsafe-inline';
    img-src
      https://web-platform.test:8444/web-bundle/resources/wbn/subresource.wbn;
    frame-src
      urn:"
>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<link rel="webbundle" href="../resources/wbn/subresource.wbn"
      resources="https://web-platform.test:8444/web-bundle/resources/wbn/fail.png" />
<link rel="webbundle" href="../resources/wbn/urn-uuid.wbn"
      resources="urn:uuid:020111b3-437a-4c5c-ae07-adb6bbffb720
                 urn:uuid:429fcc4e-0696-4bad-b099-ee9175f023ae" />
<script>
  function expect_violation() {
    return new Promise((resolve) => {
      document.addEventListener('securitypolicyviolation', (e) => {
        e.stopPropagation();
        resolve(e);
      }, {once: true});
    });
  }

  promise_test(async () => {
    const p = expect_violation();
    const img = document.createElement('img');
    img.src = 'https://web-platform.test:8444/web-bundle/resources/wbn/fail.png';
    document.body.appendChild(img);
    const e = await p;
    assert_equals(e.blockedURI, img.src);
  }, 'URL matching of CSP should be done based on the subresource URL, ' +
     'not on the bundle URL, when the subresource URL is HTTPS URL.');

  promise_test(async () => {
    const urn_uuid = 'urn:uuid:020111b3-437a-4c5c-ae07-adb6bbffb720';
    const p = expect_violation();
    const script = document.createElement('script');
    script.src = urn_uuid;
    document.body.appendChild(script);
    const e = await p;
    // Use prefix check because browsers may strip cross-origin blockedURI.
    assert_true(urn_uuid.startsWith(e.blockedURI));
  }, 'URL matching of script-src CSP should be done based on the bundle URL ' +
     'when the subresource URL is urn:uuid URL.');

  promise_test(async () => {
    const bundle_url = 'https://web-platform.test:8444/web-bundle/resources/wbn/urn-uuid.wbn';
    const urn_uuid = 'urn:uuid:429fcc4e-0696-4bad-b099-ee9175f023ae';
    const p = expect_violation();
    const iframe = document.createElement('iframe');
    iframe.src = urn_uuid;
    document.body.appendChild(iframe);
    const e = await p;
    // Currently Chromium is reporting the bundle URL for 'frame-src'. This
    // behavior is different from subresources. It may be better to change the
    // behavior to report the frame URL. But reporting anything at all for
    // 'frame-src' violations is problematic
    // (https://github.com/web-platform-tests/wpt/issues/27384). And we may
    // consider dropping reporting completely for frame-src
    // (https://crrev.com/c/2799711). So we keep this this behavior as is for
    // now.
    assert_equals(e.blockedURI, bundle_url);
  }, 'URL matching of frame-src CSP should be done based on the bundle URL ' +
     'when the frame URL is urn:uuid URL.');
</script>
</body>
